Removed ancient bug in multiple relationalBones
The bug occured, when RelationalMultiSelectionBone.unserialize() was called multiple times. This is the case in EditWidget.setData() on the bottom, when hashArgs are provided. RelationalBones loose their entires in such a case, especially when the bone itself is never referenced. This fix prevents it now, and simply ignores the bone until it is really unserialized.