Working draft: Requesting views on demand
This little enhancement makes it possible to provide a function in the adminInfo for List-modules that is called when the module or a view of the module is
activated, to load more views afterwards. The reason for doing this is in huger projects that have many different, partly generated views that may take too much time for rendering at startup. So views can be loaded on-demand from the server by calling a user-defined function that returns the views and makes them accessible.