Refactored components.Button for better integration
Using this change, you can also use the same button as tag now in HTML code and set its text and icon attributes. The button is also automatically bound to the function onHelloBtnClick, if it exists and no other callback was defined.

```python
class Bla(html5.Div):
	def __init__():
		super().__init__("""
			<button [name]="helloBtn" class="btn-hello" text="Hello" icon="icons-hello"></button>""")

	def onHelloBtnClick():
		html5.ext.Alert("Hello World")
```

The manual Button usages as in all cases in Vi are compatible.